Avoid -Wconversion warning in detail/md5.hpp

+++ include/boost/uuid/detail/md5.hpp
@@ -260,7 +260,7 @@ private:
         saved_lo = ctx->lo;
         if ((ctx->lo = (saved_lo + size) & 0x1fffffff) < saved_lo)
             ctx->hi++;
-        ctx->hi += size >> 29;
+        ctx->hi += static_cast<std::uint32_t>( size >> 29 );
 
         used = saved_lo & 0x3f;
